pyproject.toml
[tool.poetry]
name = "most-recent-file"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
argcomplete = "^1.12.3"
GitPython = "^3.1.18"

[tool.poetry.dev-dependencies]
pytest = "^6.0"
black = "^21.6b0"
ipython = "^7.25.0"
mypy = "^0.910"
pylint = "^2.9.1"

[tool.poetry.scripts]
most-recent-file = 'most_recent_file:cli'

[tool.pytest.ini_options]
log_cli = true
log_cli_level = "info"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
